Cast unsigned argument to int in call to isdigit to silence gcc warning.
git-svn-id: https://svn.parrot.org/parrot/trunk@895 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ chartypes/unicode.c
@@ -24,7 +24,7 @@ unicode_transcode_to(const char *to) {
 
 static BOOLVAL
 unicode_is_digit(UINTVAL c) {
-    return (BOOLVAL)(isdigit(c) ? 1 : 0); /* FIXME - Other code points are also digits */
+    return (BOOLVAL)(isdigit((int) c) ? 1 : 0); /* FIXME - Other code points are also digits */
 }
 
 static INTVAL
